requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
evaluate>=0.4.0

# API clients
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0

# Evaluation and metrics
scikit-learn>=1.0.0
scipy>=1.7.0
nltk>=3.8.0
textstat>=0.7.3

# Bias detection tools
# perspective-api-client>=0.1.0  # Not available on PyPI, install separately if needed

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0
rich>=13.0.0
requests>=2.28.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.0.0
nbformat>=5.7.0